difficulties with votes
